requirements.txt
streamlit
pandas
pillow
matplotlib
seaborn
cloudinary